Fixed crash, and the duration being too short.
If the user when to a page beyond the first one, and then created a new song, the program would crash as the measure would be beyond the vector length. Added a check. Made the duration no longer hard coded to 10.0 seconds, instead its the number of notes multiplied by the beats per second. Cleaned up some code.